Exception delete

+++ src/main/java/com/joaovictor/energygenius/services/UserService.java
@@ -30,7 +30,12 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+	        if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+	        repository.deleteById(id);
+	    }catch (ResourceNotFoundException e){
+	        throw new ResourceNotFoundException(id);
+	    }
 	}
 	
 	public User update(Long id, User obj) {
